Refactor qr.qr_parse -> qr.parse
